+++ lib/doorkeeper/models/access_token_mixin.rb
@@ -25,11 +25,6 @@ module Doorkeeper
 
       attr_writer :use_refresh_token
 
-      if respond_to?(:attr_accessible)
-        attr_accessible :application_id, :resource_owner_id, :expires_in,
-                        :scopes, :use_refresh_token
-      end
-
       before_validation :generate_token, on: :create
       before_validation :generate_refresh_token,
                         on: :create,