+++ NEWS.md
@@ -4,7 +4,9 @@ User-visible changes worth mentioning.
 
 ---
 ## 4.0.0.rc3
+
 - Toughen parameters filter with exact match
+- Drop `attr_accessible` from models
 
 ### Backward incompatible changes
 - Force all timezones to use UTC to prevent comparison issues.